fix make more readable

+++ app/resources/questions/questions.py
@@ -24,14 +24,11 @@ def home():
     """show all questions"""
     try:
         connection=database_connection("development")
-        try:
-            cursor = question.fetch_all_question(connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            questions_list=cursor.fetchall()
-            if questions_list:
-                return jsonify({"questions": questions_list}),200
-            abort(404)
-        except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),400
+        cursor = question.fetch_all_question(connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+        questions_list=cursor.fetchall()
+        if questions_list:
+            return jsonify({"questions": questions_list}),200
+        abort(404)
         connection.close()
     except (Exception, psycopg2.DatabaseError) as e:
         return jsonify({"error": "request error please check your request body"}),400
@@ -43,14 +40,12 @@ def get_question(question_id):
     current_user = get_jwt_identity()
     try:
         connection=database_connection("development")
-        try:
-            cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-            new_question=cursor.fetchall()
-            if new_question:
-                return jsonify({"question":new_question[0]}),200
-            return jsonify({"error":"question not found"})
-        except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),400
+
+        cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+        new_question=cursor.fetchall()
+        if new_question:
+            return jsonify({"question":new_question[0]}),200
+        return jsonify({"error":"question not found"})
         abort(404)
         connection.close()
     except Exception as e:
@@ -63,26 +58,18 @@ def post_question():
     try:
         current_user = get_jwt_identity()
         connection = database_connection("development")
-        if not request.json:
-             return jsonify({"error":"your the content you send is empty"}),400
-        if not 'title' in request.json and type(request.json['title']) != str:
-             return jsonify({"error":"key error 'title' in your request"})
-        if not "description" in request.json:
-            return jsonify({"error":"key error 'description' in your request"})
+        if not request.json or  not 'title' in request.json or not "description" in request.json :
+             return jsonify({"error":"Bad request format"}),400
         userid=current_user[0]["userid"]
         title= request.json['title']
         description = request.json['description']
         time_created = datetime.utcnow()
-        try:
-            if title.strip() and description.strip():
-                question.add_question(title,description,time_created,userid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                return jsonify({"info":"question asked"}),201
-            return jsonify({"error":"you have an empty key value in you body"}),400
-        except (Exception, psycopg2.DatabaseError) as error:
-            return jsonify({"error":str(error)}),400#jsonify({"error": "request error please check your request body"}),400
-
+        if title.strip() and description.strip():
+            question.add_question(title,description,time_created,userid,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+            return jsonify({"success":"question asked"}),201
+        return jsonify({"error":"bad request format"}),400
     except (Exception, psycopg2.DatabaseError) as e:
-        return jsonify({"error": "request error please check your request body"}),400
+        return jsonify({"error": "Bad request format"}),400
 
 @QUESTION.route('/api/v1/update_question/<int:question_id>', methods=["PUT"])
 @jwt_required
@@ -90,29 +77,27 @@ def update_question(question_id):
     '''edit question'''
     current_user = get_jwt_identity()
     connection=database_connection("development")
-    if not request.json:
-         return jsonify({"error":"your the content you send is empty"}),400
-    if not 'title' in request.json and type(request.json['title']) != str:
-         return jsonify({"error":"key error 'title' in your request"}),400
-    if not 'description' in request.json and type(request.json['description']) is not str:
-          return jsonify({"error":"key error 'description' in your request"}),400
+    if not request.json or not 'title' in request.json \
+    or not 'description' in request.json:
+        return jsonify({"error":"bad request format"}),400
     try:
         cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
         current_question=cursor.fetchall()
         if current_question:
-            #if current_user[0]["userid"]==current_question[0]["userid"]:
+            userid=current_user[0]["userid"]
             title = request.json.get('title')
             description= request.json.get('description')
             if title.strip() and description.strip():
-                question.update_question(title, description, question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
-                new_question=cursor.fetchall()
-                return jsonify({"question":new_question}),200
-            return jsonify({"error":"you have an empty key value in you body"}),400
-            #abort(403)
+                if userid == current_question[0]["userid"]:
+                    question.update_question(title, description, question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                    cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
+                    new_question=cursor.fetchall()
+                    return jsonify({"question":new_question}),200
+                return jsonify({"Warning":"permission denied"})
+            return jsonify({"error":"bad request format"}),400
         return jsonify({"error":"no question found"}),404
     except Exception as error:
-        return jsonify({"error":str(error)}),400
+        return jsonify({"error":"sorry your request cannot be processed"}),400
 
 @QUESTION.route('/api/v1/delete_question/<int:question_id>', methods=["DELETE"])
 @jwt_required
@@ -124,12 +109,12 @@ def delete_question(question_id):
         cursor=question.search_question_by_questionid(question_id,connection.cursor(cursor_factory = psycopg2.extras.RealDictCursor))
         delete_question=cursor.fetchall()
         if delete_question:
-            #if current_user[0]["userid"] == delete_question[0]["userid"]:
+            if current_user[0]["userid"] == delete_question[0]["userid"]:
                 if question.delete_question(question_id,connection.cursor()):
-                    return jsonify({"result": " question deleted"}),200
+                    return jsonify({"success": " question deleted"}),200
                 return jsonify({"error": "no question found"})
-            #abort(403)
+            return jsonify({"Warning":"permission denied"})
         return jsonify({"error":"no question found"}),404
     except (Exception,psycopg2.DatabaseError) as error:
-        return jsonify({"error": "request error please check your request body"}),400
+        return jsonify({"error": str(error)}),400
     connection.close()
